refactor: Move Footer to separate file
