+++ pages/_app.tsx
@@ -25,89 +25,10 @@ import { useRouter } from "next/router";
 import React, { ReactNode } from "react";
 import { FaGithub, FaLinkedin, FaTwitter } from "react-icons/fa";
 import JDLink from "../components/basics/JDLink";
+import Footer from "../components/Footer";
 import Logo from "../components/Logo";
 import Navbar from "../components/Navbar";
 import { supabase } from "../lib/initSupabase";
-
-const SocialButton = ({
-  children,
-  label,
-  href,
-}: {
-  children: ReactNode;
-  label: string;
-  href: string;
-}) => {
-  return (
-    <chakra.button
-      bg={useColorModeValue("blackAlpha.100", "whiteAlpha.100")}
-      rounded={"full"}
-      w={8}
-      h={8}
-      cursor={"pointer"}
-      as={"a"}
-      href={href}
-      display={"inline-flex"}
-      alignItems={"center"}
-      justifyContent={"center"}
-      transition={"background 0.3s ease"}
-      _hover={{ color: "#e53e3e" }}
-      color="gray.600"
-    >
-      <VisuallyHidden>{label}</VisuallyHidden>
-      {children}
-    </chakra.button>
-  );
-};
-
-function JDFooter() {
-  return (
-    <Center>
-      <Container
-        as={Stack}
-        // mt={5}
-        py={5}
-        direction={{ base: "column", md: "row" }}
-        justify={{ base: "center", md: "space-between" }}
-        align={{ base: "center", md: "center" }}
-        width="100vw"
-        maxWidth="1260px"
-        color="gray.400"
-      >
-        <Text>© 2021 JustDecide. All rights reserved.</Text>
-        <Text>
-          Made by{" "}
-          <JDLink href={"https://github.com/sponsors/bijancn"} color="gray.600">
-            Bijan Chokoufe Nejad
-          </JDLink>
-        </Text>
-        <JDLink href="/terms-and-privacy" color="gray.600">
-          Terms & Privacy
-        </JDLink>
-        <Stack direction={"row"} spacing={6}>
-          <SocialButton
-            label={"Twitter"}
-            href={"https://twitter.com/JustDecideApp"}
-          >
-            <FaTwitter />
-          </SocialButton>
-          <SocialButton
-            label={"Github"}
-            href={"https://github.com/bijancn/justdecide"}
-          >
-            <FaGithub />
-          </SocialButton>
-          <SocialButton
-            label={"LinkedIn"}
-            href="https://www.linkedin.com/in/bijanchokoufe/"
-          >
-            <FaLinkedin />
-          </SocialButton>
-        </Stack>
-      </Container>
-    </Center>
-  );
-}
 
 const theme = extendTheme({
   fonts: {
@@ -145,7 +66,7 @@ function InnerApp({ Component, pageProps }) {
               </Container>
             </Center>
           )}
-          <JDFooter />
+          <Footer />
         </VStack>
       </ChakraProvider>
     </div>
